move fonts to the vendor folder; move useValidation hook to the hooks folder.
