+++ packages/client/src/components/Form/hooks.ts
@@ -1,11 +1,5 @@
 import { useState } from 'react';
 
-export type ValidationData = {
-  field: string;
-  isValid: boolean;
-  text?: string;
-};
-
 export function useForm(defaultValues?: Record<string, string | undefined>) {
   const [values, setValues] = useState<Record<string, string | undefined>>(
     defaultValues ?? {}
@@ -18,31 +12,4 @@ export function useForm(defaultValues?: Record<string, string | undefined>) {
   }
 
   return { values, handleInputChange };
-}
-
-export function useValidation(
-  validations: { field: string; validation: any }[]
-) {
-  const [validationData, setValidationData] = useState<ValidationData[]>([]);
-
-  function validateForm(values: Record<string, string | undefined>) {
-    const newValidationData: ValidationData[] = [];
-
-    for (const validation of validations) {
-      newValidationData.push({
-        ...validation.validation(values[validation.field]),
-        field: validation.field,
-      });
-    }
-
-    setValidationData(newValidationData);
-
-    return !newValidationData.some(v => v.isValid === false);
-  }
-
-  function clearFieldValidation(field: string) {
-    setValidationData(validationData.filter(d => d.field !== field));
-  }
-
-  return { validationData, validateForm, clearFieldValidation };
 }
