+++ packages/client/src/index.css
@@ -6,18 +6,20 @@
 
 @font-face {
   font-family: 'Inter';
-  src: local('Inter-Regular'), url('./fonts/Inter-Regular.woff') format('woff');
+  src: local('Inter-Regular'),
+    url('./vendor/fonts/Inter-Regular.woff') format('woff');
   font-weight: 400;
 }
 
 @font-face {
   font-family: 'Inter';
-  src: local('Inter-Medium'), url('./fonts/Inter-Medium.woff') format('woff');
+  src: local('Inter-Medium'),
+    url('./vendor/fonts/Inter-Medium.woff') format('woff');
   font-weight: 500;
 }
 
 @font-face {
   font-family: 'Inter';
-  src: local('Inter-Bold'), url('./fonts/Inter-Bold.woff') format('woff');
+  src: local('Inter-Bold'), url('./vendor/fonts/Inter-Bold.woff') format('woff');
   font-weight: 700;
 }
